requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
colorlog==6.9.0
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
geographiclib==2.0
geopy==2.4.1
idna==3.10
inflection==0.5.1
mssql-django==1.5
packaging==24.1
PyJWT==2.9.0
pyodbc==5.1.0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
setuptools==75.1.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wheel==0.44.0
